requirements.txt
# ========== 核心模組 ==========
python-dotenv==1.1.0

# ========== 向量處理 ==========
faiss-cpu==1.10.0
numpy==2.2.4
sentence-transformers==2.7.0  # ✅ HuggingFace Embeddings

# ========== PDF 處理 ==========
PyMuPDF==1.25.5  # fitz

# ========== 進度顯示 ==========
tqdm==4.67.1

# ========== UI 與展示 ==========
streamlit==1.44.1
matplotlib==3.10.1
